+++ mapbox-maps-ios/CHANGELOG.md
@@ -57,6 +57,7 @@ Mapbox welcomes participation and contributions from everyone.
 * Experimental API `MapboxMap/setMemoryBudget` was renamed to `MapboxMaps/setTileCacheBudget` and promoted to stable.
 * Location consumer methods have been renamed to align with Swift API Design Guidelines. Use `addLocationConsumer(_:)` and `removeLocationConsumer(_:)` rather than `addLocationConsumer(newConsumer:)` and `removeLocationConsumer(consumer:)`.
 * `SourceType` and `LayerType` are now structs with static variables instead of enums
+* Expose new 3D Lights API: `AmbientLight` and `DirectionalLight`.
 * Extend `Layer` protocol with `visibility` property.
 
 ## 10.12.0 - March 22, 2023